+++ test/EasyPost/TrackerTest.php
@@ -5,6 +5,7 @@ namespace EasyPost\Test;
 use VCR\VCR;
 use EasyPost\Tracker;
 use EasyPost\EasyPost;
+use EasyPost\Test\Fixture;
 
 EasyPost::setApiKey(getenv('EASYPOST_TEST_API_KEY'));
 
@@ -32,7 +33,7 @@ class TrackerTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * Test the creation of a Tracker
+     * Test creating a Tracker
      *
      * @return void
      */
@@ -45,7 +46,6 @@ class TrackerTest extends \PHPUnit\Framework\TestCase
         ]);
 
         $this->assertInstanceOf('\EasyPost\Tracker', $tracker);
-        $this->assertIsString($tracker->id);
         $this->assertStringMatchesFormat('trk_%s', $tracker->id);
         $this->assertEquals($tracker->status, 'pre_transit');
 
@@ -54,7 +54,7 @@ class TrackerTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * Test the retrieval of a Tracker
+     * Test retrieving a Tracker
      *
      * @param Tracker $tracker
      * @return void
@@ -72,7 +72,7 @@ class TrackerTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * Test the retrieval of all trackers
+     * Test retrieving all trackers
      *
      * @return void
      */
@@ -80,20 +80,16 @@ class TrackerTest extends \PHPUnit\Framework\TestCase
     {
         VCR::insertCassette('trackers/all.yml');
 
-        $page_size = 5;
-
         $trackers = Tracker::all([
-            'page_size' => $page_size,
+            'page_size' => Fixture::page_size(),
         ]);
 
         $trackers_array = $trackers['trackers'];
-        $first_tracker = $trackers['trackers'][0];
 
-        $this->assertIsArray($trackers_array);
-        $this->assertCount($page_size, $trackers_array);
-        $this->assertInstanceOf('\EasyPost\Tracker', $first_tracker);
-        $this->assertIsString($first_tracker->id);
-        $this->assertStringMatchesFormat('trk_%s', $first_tracker->id);
+        $this->assertLessThanOrEqual($trackers_array, Fixture::page_size());
+        foreach ($trackers_array as $tracker) {
+            $this->assertInstanceOf('\EasyPost\Tracker', $tracker);
+        }
     }
 
     /**
@@ -105,14 +101,13 @@ class TrackerTest extends \PHPUnit\Framework\TestCase
     {
         VCR::insertCassette('trackers/createList.yml');
 
-        Tracker::create_list([
+        $response = Tracker::create_list([
             ["tracking_code" => "EZ1000000001"],
             ["tracking_code" => "EZ1000000002"],
             ["tracking_code" => "EZ1000000003"],
-            ["tracking_code" => "EZ1000000004"],
         ]);
 
-        // This endpoint/method does not return anything, just make sure the request doesn't fail
-        $this->expectNotToPerformAssertions();
+        // This endpoint returns nothing so we assert the function returns true
+        $this->assertEquals(true, $response);
     }
 }