+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Set a default timeout of 30 seconds for connections and 60 seconds for requests
 * Fixes the tracker `create_list` method's encoding to properly send through the parameters
 * Adds the PHP version used to the `user-agent` header
+* Removed unusable `all()` method on the Parcel object
 
 ### 4.0.2 2021-10-20